requirements.txt
streamlit
pdfplumber
python-docx
spacy
scikit-learn
openai==0.28
